Add files via upload
As found by HKZ on Mastodon (@hkz@chinwag.org) the "WHITE" variable in the original Scan converter repo (https://github.com/guruthree/mac-se-video-converter) was defined as 0x7FFF. This resulted in a distinctly yellow display. Correcting this value to 0xFFFE fixed this issue, resulting in a nice crisp monochrome display.